changed name of newTool to exampleTool and updated documentation
