+++ cmd/exampleTool/testdata/test.bed
