+++ cmd/exampleTool/testdata/test.fasta
